refactor: remove utils from token-bridge-sdk
