+++ packages/arb-token-bridge-ui/src/components/TransferPanel/TokenImportDialog.tsx
@@ -7,17 +7,17 @@ import Tippy from '@tippyjs/react'
 import Image from 'next/image'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useLatest } from 'react-use'
-import { ERC20BridgeToken, getL1TokenData } from 'token-bridge-sdk'
+import { ERC20BridgeToken } from 'token-bridge-sdk'
 
 import { useERC20L1Address } from '../../hooks/useERC20L1Address'
 import { useNetworksAndSigners } from '../../hooks/useNetworksAndSigners'
 import { useActions, useAppState } from '../../state'
 import { getExplorerUrl } from '../../util/networks'
+import { getL1TokenData } from '../../util/TokenUtils'
 import { Loader } from '../common/atoms/Loader'
 import { Dialog, UseDialogProps } from '../common/Dialog'
 import { SafeImage } from '../common/SafeImage'
 import GrumpyCat from '@/images/grumpy-cat.webp'
-
 import {
   toERC20BridgeToken,
   useTokensFromLists,
